requirements.txt
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
decorator==4.4.2
Deprecated==1.2.13
ffmpeg-python==0.2.0
Flask==2.0.2
future==0.18.2
google-api-core==2.2.2
google-auth==2.3.3
google-cloud-container==2.10.1
google-cloud-core==2.2.1
google-cloud-speech==2.11.1
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
grpc-google-iam-v1==0.12.3
grpcio==1.42.0
grpcio-status==1.42.0
idna==3.3
imageio==2.13.0
imageio-ffmpeg==0.4.5
itsdangerous==2.0.1
jellyfish==0.8.9
Jinja2==3.0.3
jsonpickle==2.0.0
libcst==0.3.23
MarkupSafe==2.0.1
moviepy==1.0.3
mypy-extensions==0.4.3
networkx==2.6.3
numpy==1.21.4
pika==1.2.0
Pillow==8.4.0
proglog==0.1.9
proto-plus==1.19.8
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyYAML==6.0
redis==4.0.2
regex==2021.11.10
requests==2.26.0
rsa==4.8
segtok==1.5.10
six==1.16.0
tabulate==0.8.9
tqdm==4.62.3
typing-inspect==0.7.1
typing_extensions==4.0.0
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.13.3
yake==0.4.8
